renamed wikiLinkRender to customLinkRender to remove name collision with WikiLinkRender
